♻️ refactor(api) [#10.11.16]: 2차 개선 - 에러 핸들러 동적 메시지 및 설정 의존성 최적화

+++ backend/api/endpoints/classify.py
@@ -24,7 +24,7 @@ async def classify_file(
     # Read file content to check size
     content = await file.read()
     if len(content) > max_file_size:
-        max_size_mb = max_file_size / (1024 * 1024)
+        max_size_mb = int(max_file_size / (1024 * 1024))
         raise localized_http_exception(
             status_code=413,
             message_key="payload_too_large",
